+++ src/Utils.ts
@@ -1,3 +1,3 @@
 export default {
-    Version: "2.2.1"
+    Version: "2.2.2"
 }